+++ posts/src/main/java/com/neaniesoft/vermilion/posts/domain/entities/AnimatedImagePreview.kt
@@ -1,0 +1,9 @@
+package com.neaniesoft.vermilion.posts.domain.entities
+
+import android.net.Uri
+
+data class AnimatedImagePreview(
+    val uri: Uri,
+    val width: Int,
+    val height: Int
+)
