+++ posts/src/main/java/com/neaniesoft/vermilion/posts/domain/entities/MarkdownText.kt
@@ -1,0 +1,8 @@
+package com.neaniesoft.vermilion.posts.domain.entities
+
+import org.commonmark.node.Document
+
+data class MarkdownText(
+    val raw: String,
+    val markdown: Document
+)
