+++ posts/src/main/java/com/neaniesoft/vermilion/posts/domain/entities/UriImage.kt
@@ -1,0 +1,9 @@
+package com.neaniesoft.vermilion.posts.domain.entities
+
+import android.net.Uri
+
+data class UriImage(
+    val uri: Uri,
+    val width: Int,
+    val height: Int
+)
